+++ src/main/java/com/bridegelabz/employeepayrollapp/repository/EmployeePayrollRepository.java
@@ -3,5 +3,5 @@ package com.bridegelabz.employeepayrollapp.repository;
 import com.bridegelabz.employeepayrollapp.entity.EmployeePayrollData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeePayrollRepository extends JpaRepository<EmployeePayrollData, Integer> {
+public interface EmployeePayrollRepository extends JpaRepository<EmployeePayrollData, Long> {
 }
